auth/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
python-dotenv
psycopg2-binary
